requirements.txt
requests
beautifulsoup4
bs4
lxml
pandas
tqdm
numpy
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
torch>=1.12.0
